feat: soporte hardware real — AHCI/SATA, ACPI shutdown/reboot, framebuffer 16bpp
- Driver AHCI/SATA completo (drivers/disk/ahci.c):
  PCI discovery, MMIO BAR5, DMA read/write (LBA48),
  IDENTIFY DEVICE, registra drives como sd0/sd1 en blkdev
- ACPI FADT: shutdown (S5 via PM1a/PM1b), reboot
  (reset register + PS/2 + triple fault fallback)
- Comandos shell: shutdown, poweroff, halt, reboot
- Framebuffer: soporte 16bpp RGB565, resolución 1024x768
- Migración completa a x86_64 (long mode, PML4, GDT64, ELF64)
- Documentación actualizada (README, TECHNICAL, CHECKLIST)

+++ include/kernel/elf.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef struct {
-    uint32_t entry;
+    uintptr_t entry;
     uint32_t program_header_offset;
     uint16_t program_header_count;
     uint16_t program_header_entry_size;
